[GEOS-8363] Creation of SQL View Layer with Parameters fails
